% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptfce.r
\name{aggregate_logpvals}
\alias{aggregate_logpvals}
\title{Aggregate logpvals by the "equidistant incremental logarithmic probability aggregation".}
\usage{
aggregate_logpvals(logpvals, d)
}
\arguments{
\item{logpvals}{vector of enhnaced -logP values correspoinding tho various thresholds}

\item{d}{delta -logP (giving the equidistant distribution in the log-space)}
}
\value{
aggregated -logP probability
}
\description{
Aggregate logpvals by the "equidistant incremental logarithmic probability aggregation".
}
