% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptfce.r
\name{Es}
\alias{Es}
\title{expected value of cluster size at threshold h in img of size V with RESEL count Rd
mainly ported from the source code of FSL (https://fsl.fmrib.ox.ac.uk/fsl/fslwiki)}
\usage{
Es(h, V, Rd)
}
\arguments{
\item{h}{image height, that is, Z score threshold}

\item{V}{Number of voxels}

\item{Rd}{Rd}
}
\value{
Expected value of cluster size
}
\description{
expected value of cluster size at threshold h in img of size V with RESEL count Rd
mainly ported from the source code of FSL (https://fsl.fmrib.ox.ac.uk/fsl/fslwiki)
}
