% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptfce.r
\name{fwe.p2z}
\alias{fwe.p2z}
\title{Get Z-score threshold for an FWER corrected p-value}
\usage{
fwe.p2z(resel_count, FWEP = 0.05)
}
\arguments{
\item{resel_count}{resel count}

\item{FWEP}{FWER corrected p-value}
}
\value{
Z-score threshold
}
\description{
https://github.com/spisakt/pTFCE
}
\details{
The p-value, corrected for multiple comparisons by controlling for family-wise error rate, is converted to a Z-score threshold.
The parameter resel_count is the number of resels (resolution elements) in the image, and can be obtained e.g by smoothest() (see examples).
}
\examples{
s=smoothest(zmap, mask)
fwe.p2z(resel_count=s$volume/s$resels, FWEP=0.05)

pTFCE=ptfe(zmap, mask)
fwe.p2z(pTFCE$number_of_resels, FWEP=0.05)
}
