% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptfce.r
\name{fwe.z2p}
\alias{fwe.z2p}
\title{Convert Z-score value to FWER corrected p-value}
\usage{
fwe.z2p(resel_count, Z)
}
\arguments{
\item{resel_count}{resel count}

\item{Z}{z-score value}
}
\value{
FWER-corrected p-value
}
\description{
https://github.com/spisakt/pTFCE
}
\details{
The Z-score value is converted to to p-value, corrected for multiple comparisons by controlling for famili-wise error rate
The parameter resel_count is the number of resels (resolution elements) in the image, and can be obtained e.g by smoothest() (see examples).
}
\examples{
s=smoothest(zmap, mask)
fwe.z2p(resel_count=s$volume/s$resels, Z=2.3)

pTFCE=ptfe(zmap, mask)
fwe.z2p(pTFCE$number_of_resels, Z=3.1)
}
