% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptfce.r
\name{ptfce}
\alias{ptfce}
\title{pTFCE main function.}
\usage{
ptfce(img, Rd, V, mask, length.out = 50, logpmin = 0,
  logpmax = -log(pnorm(max(img), lower.tail = F)))
}
\arguments{
\item{img}{Nifti image to enhance ("nifti" class from "oro.nifti" package)}

\item{Rd}{Resel count (as output by FSL smoothest)}

\item{V}{Number of voixels in mask}

\item{mask}{Mask}

\item{length.out}{Number of thresholds}

\item{logpmin}{min threshold}

\item{logpmax}{max threshold}
}
\value{
TFCE object
}
\description{
The threshold-free cluster enhancement (TFCE) approach integrates cluster information into voxel-wise statistical inference to enhance detectability of neuroimaging signal. Despite the significantly increased sensitivity, the application of TFCE is limited by several factors: (i) generalization to data structures, like brain network connectivity data is not trivial, (ii) TFCE values are in an arbitrary unit, therefore, P-values can only be obtained by a computationally demanding permutation-test.
Here, we introduce a probabilistic approach for TFCE (pTFCE), that gives a simple general framework for topology-based belief boosting.
The core of pTFCE is a conditional probability, calculated based on Bayes' rule, from the probability of voxel intensity and the threshold-wise likelihood function of the measured cluster size. We provide an estimation of these distributions based on Gaussian Random Field (GRF) theory. The conditional probabilities are then aggregated across cluster-forming thresholds by a novel incremental aggregation method. Our approach is validated on simulated and real fMRI data.
pTFCE is shown to be more robust to various ground truth shapes and provides a stricter control over cluster "leaking" than TFCE and, in the most realistic cases, further improves its sensitivity. Correction for multiple comparison can be trivially performed on the enhanced P-values, without the need for permutation testing, thus pTFCE is well-suitable for the improvement of statistical inference in any neuroimaging workflow.
}
\examples{
Z=readNIfTI("Zmap.nii.gz");
MASK=readNIfTI("mask.nii.gz");
smooth=read.table("smoothness.txt");
ptfce(Z, V=smooth[2,2], Rd = smooth[1,2]*smooth[2,2], mask = MASK, length.out = 100);
}
